Made several changes to code:
HAMON.py - Fixed so that HAMON can be executed as stand alone program
(must chmod +x to make it work). Also fixed so that all input arguments
are read inside the main program, not outside. Only modules are loaded
there.

src/__init__.py - Fixed an init file so that src/ folder is recognized
by python as a module folder. In all source files, I have also updated
so that the the folder infrastructure is adopted when modules are
loaded.

src/differential_evolution.py - Fixed so that the config file variables
are passed as an input to the constructor of both the single and multi
objective classes being defined in this file.

src/genetic_algorithm.py - same as differential_evolution.py

src/populations.py - Fixed how modules are being loaded.

src/read_input_arguments.py - Changed to more robust read command, which
will warn the user if not enough arguments are being supplied, and also
exit the code with an appropriate error warning.